update 3.9.2.1
Added a way to set what mod types you want for your checked mods such as pseudo, explicit, implicit, crafted that can be set in your config file.
If you do not wish to replace your whole config file, replace the "checked" list in the config file with the one in the latest config file.
Fixed a bug where if you did not check off the #% total elemental resistance mod, the mod would re-appear in the next item it parsed even though it did not have that mod.